improve Partition trait
Earlier, the user had to implement the trait such that partitioning is possible. An ugly side effect is that an external trait (Partition) had to be implemented on an external type (LocalMap) which is not possible; One would have to manually re-implement each trait using a newtype pattern, which leads to a lot of unnecessary boilerplate code.

The new solution extends the Partition trait such that you can set and get a partitioning algorithm. That way, the user only needs to call the setter function to provide an algorithm, and it avoids all of the boilerplate code.